requirements-cs5914.txt
pyautogen
requests
streamlit
openai
toml
pandas
tabulate
python-dotenv
httpx
asyncio~=3.4.3
pytest
fastapi
uvicorn
gunicorn
jupyter

